+++ src/lib/indicator/algorithm/sto.js
@@ -75,7 +75,7 @@ export default function() {
 		return newData;
 	}
 	calculator.undefinedLength = function() {
-		return Math.min(windowSize + kWindowSize, dWindowSize);
+		return windowSize + kWindowSize + dWindowSize;
 	};
 	calculator.windowSize = function(x) {
 		if (!arguments.length) {