+++ src/lib/tooltip/StochasticTooltip.jsx
@@ -65,7 +65,10 @@ StochasticTooltip.propTypes = {
 	fontFamily: PropTypes.string,
 	fontSize: PropTypes.number,
 	onClick: PropTypes.func,
-	calculator: PropTypes.func.isRequired,
+	calculator: PropTypes.oneOfType([
+		PropTypes.func,
+		PropTypes.object,
+	]).isRequired,
 	displayFormat: PropTypes.func.isRequired,
 	children: PropTypes.node.isRequired,
 };